backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
redis==5.0.1
celery==5.3.4
stripe==7.8.0
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
gunicorn==21.2.0 